LayoutTests/media/track/track-webvtt-tc021-valign-expected.txt

Tests cue vertical alignment (direction) from settings.


*** Testing text track 0
EXPECTED (cues.length == '3') OK
EXPECTED (cues[0].vertical == 'rl') OK
EXPECTED (cues[0].align == 'middle') OK
EXPECTED (cues[0].position == '50') OK
EXPECTED (cues[1].vertical == 'lr') OK
EXPECTED (cues[1].align == 'middle') OK
EXPECTED (cues[1].position == '50') OK
EXPECTED (cues[2].vertical == 'rl') OK
EXPECTED (cues[2].align == 'start') OK
EXPECTED (cues[2].position == '0') OK

*** Testing text track 1
EXPECTED (cues.length == '3') OK
EXPECTED (cues[0].vertical == 'rl') OK
EXPECTED (cues[0].align == 'middle') OK
EXPECTED (cues[0].position == '50') OK
EXPECTED (cues[1].vertical == 'lr') OK
EXPECTED (cues[1].align == 'middle') OK
EXPECTED (cues[1].position == '50') OK
EXPECTED (cues[2].vertical == 'rl') OK
EXPECTED (cues[2].align == 'start') OK
EXPECTED (cues[2].position == '0') OK

*** Testing text track 2
EXPECTED (cues.length == '3') OK
EXPECTED (cues[0].vertical == '') OK
EXPECTED (cues[1].vertical == '') OK
EXPECTED (cues[2].vertical == '') OK
END OF TEST
